added authentication and authorization

+++ webstore/Services/Basket/Basket.API/Controllers/BasketController.cs
@@ -1,15 +1,17 @@
+using System.Security.Claims;
 using AutoMapper;
 using Basket.API.Entities;
-using Basket.API.Repositories;
-using Grpc.Core;
-using Microsoft.AspNetCore.Mvc;
 using Basket.API.GrpcServices;
+using Basket.API.Repositories;
 using EventBus.Messages.Events;
+using Grpc.Core;
 using MassTransit;
-
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Basket.API.Controllers;
 
+[Authorize(Roles = "Buyer")]
 [ApiController]
 [Route("api/v1/[controller]")]
 public class BasketController : ControllerBase
@@ -20,31 +22,42 @@ public class BasketController : ControllerBase
     private readonly IMapper _mapper;
     private readonly IPublishEndpoint _publishEndpoint;
 
-
-    public BasketController(IBasketRepository repository, CouponGrpcService couponGrpcService,
-        ILogger<BasketController> logger, IMapper mapper, IPublishEndpoint publishEndpoint)
+    public BasketController(
+        IBasketRepository repository, 
+        CouponGrpcService couponGrpcService, 
+        ILogger<BasketController> logger,
+        IMapper mapper,
+        IPublishEndpoint publishEndpoint)
     {
         _repository = repository ?? throw new ArgumentNullException(nameof(repository));
         _couponGrpcService = couponGrpcService ?? throw new ArgumentNullException(nameof(couponGrpcService));
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-        _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
-        _publishEndpoint = publishEndpoint ?? throw new ArgumentNullException(nameof(publishEndpoint));
+        _mapper = mapper ?? throw new ArgumentNullException(nameof(logger));
+        _publishEndpoint = publishEndpoint ?? throw new ArgumentNullException(nameof(logger));
     }
 
     [HttpGet("{username}")]
     [ProducesResponseType(typeof(ShoppingCart), StatusCodes.Status200OK)]
     public async Task<ActionResult<ShoppingCart>> GetBasket(string username)
     {
+        if (User.FindFirst(ClaimTypes.Name)?.Value != username)
+        {
+            return Forbid();
+        }
+        
         var basket = await _repository.GetBasket(username);
-
-        return Ok(basket);
+        return Ok(basket ?? new ShoppingCart(username));
     }
 
     [HttpPut]
     [ProducesResponseType(typeof(ShoppingCart), StatusCodes.Status200OK)]
-
     public async Task<ActionResult<ShoppingCart>> UpdateBasket([FromBody] ShoppingCart basket)
     {
+        if (User.FindFirstValue(ClaimTypes.Name) != basket.Username)
+        {
+            return Forbid();
+        }
+        
         foreach (var item in basket.Items)
         {
             try
@@ -52,29 +65,38 @@ public class BasketController : ControllerBase
                 var coupon = await _couponGrpcService.GetDiscount(item.ProductName);
                 item.Price -= coupon.Amount;
             }
-            catch (Exception e)
+            catch (RpcException e)
             {
                 _logger.LogInformation("Error while retrieving coupon for item {ProductName}: {message}", item.ProductName, e.Message);
             }
         }
+        
         return Ok(await _repository.UpdateBasket(basket));
     }
 
     [HttpDelete("{username}")]
     [ProducesResponseType(typeof(void), StatusCodes.Status200OK)]
-
-    public async Task<ActionResult<ShoppingCart>> DeleteBasket(string username)
+    public async Task<IActionResult> DeleteBasket(string username)
     {
+        if (User.FindFirstValue(ClaimTypes.Name) != username)
+        {
+            return Forbid();
+        }
+        
         await _repository.DeleteBasket(username);
         return Ok();
     }
+
     [Route("[action]")]
     [HttpPost]
     [ProducesResponseType(typeof(void), StatusCodes.Status202Accepted)]
     [ProducesResponseType(typeof(void), StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> Checkout([FromBody] BasketCheckout basketCheckout)
     {
-        
+        if (User.FindFirstValue(ClaimTypes.Name) != basketCheckout.BuyerUsername)
+        {
+            return Forbid();
+        }
 
         var basket = await _repository.GetBasket(basketCheckout.BuyerUsername);
         if (basket is null)
@@ -89,6 +111,4 @@ public class BasketController : ControllerBase
         
         return Accepted();
     }
-
-    
 }
